refactor(sdk): convert to esm/nodenext
