+++ packages/galileo-sdk/project.json
@@ -85,6 +85,13 @@
         "command": "pnpm exec projen eslint",
         "cwd": "packages/galileo-sdk"
       }
+    },
+    "compile:cjs": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen compile:cjs",
+        "cwd": "packages/galileo-sdk"
+      }
     }
   },
   "//": "~~ Generated by projen. To modify, edit .projenrc.js and run \"npx projen\"."
