+++ packages/galileo-sdk/src/chat/dynamodb/index.ts
@@ -1,4 +1,4 @@
 /*! Copyright [Amazon.com](http://amazon.com/), Inc. or its affiliates. All Rights Reserved.
 PDX-License-Identifier: Apache-2.0 */
-export * as lib from './lib';
-export * from './message-history';
+export * as lib from './lib/index.js';
+export * from './message-history.js';
