+++ packages/galileo-sdk/src/chat/dynamodb/lib/index.ts
@@ -1,6 +1,6 @@
 /*! Copyright [Amazon.com](http://amazon.com/), Inc. or its affiliates. All Rights Reserved.
 PDX-License-Identifier: Apache-2.0 */
-export * as Chat from './chat';
-export * as Messages from './messages';
-export * as Sources from './sources';
-export * from './util';
+export * as Chat from './chat.js';
+export * as Messages from './messages.js';
+export * as Sources from './sources.js';
+export * from './util.js';
