+++ packages/galileo-sdk/src/chat/memory.ts
@@ -1,9 +1,9 @@
 /*! Copyright [Amazon.com](http://amazon.com/), Inc. or its affiliates. All Rights Reserved.
 PDX-License-Identifier: Apache-2.0 */
 import { BufferWindowMemory, BufferWindowMemoryInput, getInputValue, BaseMemory } from 'langchain/memory';
-import { DDBChatMessage, DDBMessageSource } from './dynamodb/lib';
-import { CreateAIChatMessageResponse, CreateHumanChatMessageResponse } from './dynamodb/lib/messages';
-import { DynamoDBChatMessageHistory } from './dynamodb/message-history';
+import { DDBChatMessage, DDBMessageSource } from './dynamodb/lib/index.js';
+import { CreateAIChatMessageResponse, CreateHumanChatMessageResponse } from './dynamodb/lib/messages.js';
+import { DynamoDBChatMessageHistory } from './dynamodb/message-history.js';
 
 export interface MemoryContext {
   readonly input: DDBChatMessage;
