+++ packages/galileo-sdk/src/common/index.ts
@@ -1,5 +1,5 @@
 /*! Copyright [Amazon.com](http://amazon.com/), Inc. or its affiliates. All Rights Reserved.
 PDX-License-Identifier: Apache-2.0 */
-export * from './logging';
-export * from './env';
-export * from './types';
+export * from './logging/index.js';
+export * from './env.js';
+export * from './types.js';
