+++ packages/galileo-sdk/src/index.ts
@@ -1,9 +1,8 @@
 /*! Copyright [Amazon.com](http://amazon.com/), Inc. or its affiliates. All Rights Reserved.
 PDX-License-Identifier: Apache-2.0 */
-export * as Chat from './chat';
-export * as Infra from './infra';
-export * as Common from './common';
-export * as Models from './models';
-export * as VectorStores from './vectorstores';
+export * as Chat from './chat/index.js';
+export * as Common from './common/index.js';
+export * as Models from './models/index.js';
+export * as VectorStores from './vectorstores/index.js';
 
-export * from './langchain/vectorstores/memory';
+export * from './langchain/vectorstores/memory.js';
