--- packages/galileo-sdk/src/infra/index.ts
@@ -1,3 +1,0 @@
-/*! Copyright [Amazon.com](http://amazon.com/), Inc. or its affiliates. All Rights Reserved.
-PDX-License-Identifier: Apache-2.0 */
-export * from '../vectorstores/pgvector/rds';
