+++ packages/galileo-sdk/src/models/index.ts
@@ -1,9 +1,9 @@
 /*! Copyright [Amazon.com](http://amazon.com/), Inc. or its affiliates. All Rights Reserved.
 PDX-License-Identifier: Apache-2.0 */
-import './env';
+import './env.js';
 
-export * from './adapter';
-export * from './env';
-export * from './types';
-export * from './inventory';
-export * as LLM from './llms';
+export * from './adapter.js';
+export * from './env.js';
+export * from './types.js';
+export * from './inventory.js';
+export * as LLM from './llms/index.js';
