+++ packages/galileo-sdk/src/models/llms/index.ts
@@ -1,3 +1,3 @@
 /*! Copyright [Amazon.com](http://amazon.com/), Inc. or its affiliates. All Rights Reserved.
 PDX-License-Identifier: Apache-2.0 */
-export * as Falcon from './falcon';
+export * as Falcon from './falcon.js';
