+++ packages/galileo-sdk/src/types/global.d.ts
@@ -1,0 +1,7 @@
+declare module "cheminfo-types" {
+  export type NumberArray = any;
+}
+
+declare module "ml-tree-similarity" {
+
+}
