+++ packages/galileo-sdk/src/vectorstores/index.ts
@@ -1,3 +1,3 @@
 /*! Copyright [Amazon.com](http://amazon.com/), Inc. or its affiliates. All Rights Reserved.
 PDX-License-Identifier: Apache-2.0 */
-export * from './pgvector';
+export * from './pgvector/index.js';
