+++ packages/galileo-sdk/tsconfig.cjs.json
@@ -1,0 +1,9 @@
+// ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "declaration": false
+  }
+}
