+++ packages/galileo-sdk/tsconfig.json
@@ -11,9 +11,10 @@
     "inlineSources": true,
     "lib": [
       "DOM",
-      "ES2021"
+      "ES2021",
+      "ES2022.Object"
     ],
-    "module": "CommonJS",
+    "module": "nodenext",
     "noEmitOnError": false,
     "noFallthroughCasesInSwitch": true,
     "noImplicitAny": true,
@@ -27,11 +28,13 @@
     "strictPropertyInitialization": true,
     "stripInternal": true,
     "target": "ES2019",
-    "skipLibCheck": true,
-    "moduleResolution": "node"
+    "moduleResolution": "nodenext"
   },
   "include": [
     "src/**/*.ts"
   ],
-  "exclude": []
+  "exclude": [
+    "**/*.spec.*",
+    "**/*.test.*"
+  ]
 }
